Move generic views into contrib
